:recycle: Extract publisher.go from progress_notification.go
