+++ progress_notification.go
@@ -12,23 +12,6 @@ import (
 	log "github.com/Sirupsen/logrus"
 )
 
-type (
-	Publisher interface {
-		Publish(topic string, msg *pubsub.PubsubMessage) (*pubsub.PublishResponse, error)
-	}
-
-	pubsubPublisher struct {
-		topicsService *pubsub.ProjectsTopicsService
-	}
-)
-
-func (pp *pubsubPublisher) Publish(topic string, msg *pubsub.PubsubMessage) (*pubsub.PublishResponse, error) {
-	req := &pubsub.PublishRequest{
-		Messages: []*pubsub.PubsubMessage{msg},
-	}
-	return pp.topicsService.Publish(topic, req).Do()
-}
-
 type Progress int
 
 const (
