+++ publisher.go
@@ -1,0 +1,22 @@
+package main
+
+import (
+	pubsub "google.golang.org/api/pubsub/v1"
+)
+
+type (
+	Publisher interface {
+		Publish(topic string, msg *pubsub.PubsubMessage) (*pubsub.PublishResponse, error)
+	}
+
+	pubsubPublisher struct {
+		topicsService *pubsub.ProjectsTopicsService
+	}
+)
+
+func (pp *pubsubPublisher) Publish(topic string, msg *pubsub.PubsubMessage) (*pubsub.PublishResponse, error) {
+	req := &pubsub.PublishRequest{
+		Messages: []*pubsub.PubsubMessage{msg},
+	}
+	return pp.topicsService.Publish(topic, req).Do()
+}
